more test changes
