+++ frontend/package.json
@@ -13,8 +13,8 @@
     "@types/react": "^17.0.0",
     "@types/react-dom": "^17.0.0",
     "@types/styled-components": "^5.1.9",
-    "firebase": "^8.6.5",
     "axios": "^0.21.1",
+    "firebase": "^8.6.5",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
     "react-router-dom": "^5.2.0",
